(maint)Restructure_Solaris
