+++ lib/facter/facts/freebsd/zpool_featurenumbers.rb
@@ -6,7 +6,7 @@ module Facts
       FACT_NAME = 'zpool_featurenumbers'
 
       def call_the_resolver
-        fact_value = Facter::Resolvers::ZPool.resolve(:zpool_featurenumbers)
+        fact_value = Facter::Resolvers::Zpool.resolve(:zpool_featurenumbers)
         Facter::ResolvedFact.new(FACT_NAME, fact_value)
       end
     end
