+++ lib/facter/facts/solaris/os/release.rb
@@ -8,9 +8,9 @@ module Facts
         ALIASES = %w[operatingsystemmajrelease operatingsystemrelease].freeze
 
         def call_the_resolver
-          full_value = Facter::Resolvers::SolarisRelease.resolve(:full)
-          major_value = Facter::Resolvers::SolarisRelease.resolve(:major)
-          minor_value = Facter::Resolvers::SolarisRelease.resolve(:minor)
+          full_value = Facter::Resolvers::Solaris::OsRelease.resolve(:full)
+          major_value = Facter::Resolvers::Solaris::OsRelease.resolve(:major)
+          minor_value = Facter::Resolvers::Solaris::OsRelease.resolve(:minor)
 
           [Facter::ResolvedFact.new(FACT_NAME, full: full_value, major: major_value, minor: minor_value),
            Facter::ResolvedFact.new(ALIASES.first, major_value, :legacy),
