+++ lib/facter/resolvers/load_averages.rb
@@ -14,6 +14,8 @@ module Facter
 
         def read_load_averages(fact_name)
           require_relative 'utils/ffi/load_averages'
+
+          log.debug('loading cpu load averages')
           @fact_list[:load_averages] = %w[1m 5m 15m].zip(Utils::Ffi::LoadAverages.read_load_averages).to_h
 
           @fact_list[fact_name]
