+++ lib/facter/resolvers/solaris/filesystems.rb
