+++ lib/facter/resolvers/solaris/ipaddress.rb
