+++ lib/facter/resolvers/solaris/networking.rb
