+++ lib/facter/resolvers/zfs.rb
