+++ spec/facter/facts/solaris/zpool_featurenumbers_spec.rb
@@ -7,13 +7,13 @@ describe Facts::Solaris::ZpoolFeaturenumbers do
     let(:zpool_featurenumbers) { '1,2,3,4,5,6,7' }
 
     before do
-      allow(Facter::Resolvers::ZPool).to \
+      allow(Facter::Resolvers::Zpool).to \
         receive(:resolve).with(:zpool_featurenumbers).and_return(zpool_featurenumbers)
     end
 
     it 'calls Facter::Resolvers::ZPool' do
       fact.call_the_resolver
-      expect(Facter::Resolvers::ZPool).to have_received(:resolve).with(:zpool_featurenumbers)
+      expect(Facter::Resolvers::Zpool).to have_received(:resolve).with(:zpool_featurenumbers)
     end
 
     it 'returns the zpool_featurenumbers fact' do
