+++ spec/facter/resolvers/solaris/filesystem_spec.rb
