+++ spec/facter/resolvers/solaris/ipaddress_spec.rb
