+++ spec/facter/resolvers/zfs_spec.rb
